requirements.txt
streamlit
openai
python-whois
